Fix heap overflow in McEliece, keccak type mismatch, POSIX defines, clean test
- Fix Classic McEliece heap-buffer-overflow: gen_permutation needs field_size
  (2^m) entries but buffer was only allocated for n entries (n < 2^m)
- Fix keccak.c: rate_bytes parameter type changed from unsigned int to size_t
  to match keccak.h declaration (fixes GCC conflicting types error on 64-bit)
- Add _POSIX_C_SOURCE=200809L for Linux (fixes CLOCK_MONOTONIC in benchmarks)
- Clean up test_kem_all.c: remove all debug code, keep only round-trip tests

+++ src/core/common/hash/keccak.c
@@ -99,7 +99,7 @@ void pqc_keccak_init(uint64_t state[25]) {
     memset(state, 0, 25 * sizeof(uint64_t));
 }
 
-void pqc_keccak_absorb(uint64_t state[25], unsigned int rate_bytes,
+void pqc_keccak_absorb(uint64_t state[25], size_t rate_bytes,
                         const uint8_t *data, size_t datalen)
 {
     size_t rate_lanes = rate_bytes / 8;
@@ -128,7 +128,7 @@ void pqc_keccak_absorb(uint64_t state[25], unsigned int rate_bytes,
     }
 }
 
-void pqc_keccak_squeeze(uint64_t state[25], unsigned int rate_bytes,
+void pqc_keccak_squeeze(uint64_t state[25], size_t rate_bytes,
                          uint8_t *out, size_t outlen)
 {
     size_t rate_lanes = rate_bytes / 8;
